refactor(db): move sqlx to infra module
