--- src/db/mod.rs
@@ -1,5 +1,0 @@
-//! Database support.
-//! The database keeps incomes to query them later.
-//! It can be used as a data source instead of reading bank CSV every time.
-
-pub mod sqlx;
