+++ src/infra/sqlx/income_repository.rs
