+++ src/infra/sqlx/payment_repository.rs
@@ -1,12 +1,12 @@
 use async_trait::async_trait;
 use sqlx::{query, QueryBuilder, Sqlite, SqlitePool};
 
-use crate::db::sqlx::criteria::SqlxCriterion;
-use crate::db::sqlx::record::IncomeRecord;
 use crate::domain::repository::PaymentRepository;
 use crate::domain::Income;
 use crate::domain::Payment;
 use crate::income::criteria::{IncomeCriteria, IncomeCriterion};
+use crate::infra::sqlx::criteria::SqlxCriterion;
+use crate::infra::sqlx::record::IncomeRecord;
 
 pub struct SqlxPaymentRepository {
     pool: SqlitePool,
