+++ src/init.rs
@@ -2,7 +2,8 @@
 
 use sqlx::SqlitePool;
 
-use crate::{config::dir::create_default_config, db::sqlx::init::initialize_db};
+use crate::config::dir::create_default_config;
+use crate::infra::sqlx::init::initialize_db;
 
 pub async fn init(db_pool: &SqlitePool, force: bool) -> anyhow::Result<()> {
     create_default_config(force)?;
