+++ src/lib.rs
@@ -7,7 +7,6 @@
 pub mod app;
 pub mod cli;
 pub mod config;
-pub mod db;
 pub mod domain;
 pub mod filter;
 pub mod income;
